I think this is redundant as LXC will assume native arch.
furthermore, ARCH is set to an appropriate value during configure().
While those using custom tarballs should specify it by themselves
(I mean, we could parse it out of the tarball filename *assuming*
the tarball contains one of the values of the arches array and even
then the only valid ones would be amd64 and x86 due to LXC not
giving a fuck about others).
I guess the final answer will be given when direct Qemu emulation
support is added.